training/requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
conda==4.7.10
conda-package-handling==1.3.11
cryptography==2.7
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libarchive-c==2.8
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.2.0
protobuf==3.10.0
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
ruamel-yaml==0.15.46
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tqdm==4.32.1
urllib3==1.24.2
Werkzeug==0.16.0
wrapt==1.11.2
